src/pyproject.toml
[project]
name = "chimpvid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "elevenlabs>=1.52.0",
    "google-genai>=1.3.0",
    "moviepy==2.1.2",
    "pillow>=9.2.0,<11.0",
    "pydub>=0.25.1",
]
